refactor(RichText): extract safeHref to a shared engine module
